pyproject.toml
[tool.poetry]
name = "document-vector-search-benchmark"
version = "0.1.0"
description = ""
authors = ["TatsuyaShirakawa <lilys1204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pyyaml = "^6.0.1"
requests = "^2.31.0"
loguru = "^0.7.2"
pysen = "^0.10.5"
mypy = "^1.6.1"
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
types-requests = "^2.31.0.10"
types-pyyaml = "^6.0.12.12"
numpy = "^1.26.1"
openai = "^1.2.0"
tqdm = "^4.66.1"
tenacity = "^8.2.3"
scikit-learn = "^1.3.2"
torch = "^2.1.0"
transformers = "^4.35.0"
pandas = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pysen]
version = "0.10"


[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
line_length = 119
py_version = "py310"
[[tool.pysen.lint.mypy_targets]]
paths = ["tools", "dvsb"]


[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
